+++ app/mailers/agilibox/application_mailer.rb
@@ -1,0 +1,2 @@
+class Agilibox::ApplicationMailer < Agilibox.parent_mailer
+end